[refactor] Refactor helper function for testing into a tests/utils module
